+++ plotly/src/common/mod.rs
@@ -186,6 +186,7 @@ pub enum PlotType {
     Scatter,
     ScatterGL,
     Scatter3D,
+    ScatterMapbox,
     ScatterPolar,
     ScatterPolarGL,
     Bar,
@@ -195,6 +196,8 @@ pub enum PlotType {
     HeatMap,
     Histogram,
     Histogram2dContour,
+    Image,
+    Mesh3D,
     Ohlc,
     Sankey,
     Surface,
@@ -1487,7 +1490,8 @@ mod tests {
 
     #[test]
     fn test_serialize_direction() {
-        // TODO: I think `Direction` would be better as a struct, with `fillcolor` and `line` attributes
+        // TODO: I think `Direction` would be better as a struct, with `fillcolor` and
+        // `line` attributes
         let inc = Direction::Increasing { line: Line::new() };
         let expected = json!({"line": {}});
         assert_eq!(to_value(inc).unwrap(), expected);