+++ plotly/src/layout/update_menu.rs
@@ -10,20 +10,22 @@ use crate::{
     Relayout, Restyle,
 };
 
-/// Sets the Plotly method to be called on click. If the `skip` method is used, the API
-/// updatemenu will function as normal but will perform no API calls and will not bind
-/// automatically to state updates. This may be used to create a component interface
-/// and attach to updatemenu events manually via JavaScript.
-///
+/// Sets the Plotly method to be called on click. If the `skip` method is used,
+/// the API updatemenu will function as normal but will perform no API calls and
+/// will not bind automatically to state updates. This may be used to create a
+/// component interface and attach to updatemenu events manually via JavaScript.
 #[derive(Serialize, Debug, Copy, Clone)]
 #[serde(rename_all = "snake_case")]
 pub enum ButtonMethod {
-    /// The restyle method should be used when modifying the data and data attributes of the graph
+    /// The restyle method should be used when modifying the data and data
+    /// attributes of the graph
     Restyle,
-    /// The relayout method should be used when modifying the layout attributes of the graph.
+    /// The relayout method should be used when modifying the layout attributes
+    /// of the graph.
     Relayout,
     Animate,
-    /// The update method should be used when modifying the data and layout sections of the graph.
+    /// The update method should be used when modifying the data and layout
+    /// sections of the graph.
     Update,
     Skip,
 }
@@ -31,37 +33,44 @@ pub enum ButtonMethod {
 #[serde_with::skip_serializing_none]
 #[derive(Serialize, Clone, Debug, FieldSetter)]
 pub struct Button {
-    /// Sets the arguments values to be passed to the Plotly method set in `method` on click.
+    /// Sets the arguments values to be passed to the Plotly method set in
+    /// `method` on click.
     args: Option<Value>,
-    /// Sets a 2nd set of `args`, these arguments values are passed to the Plotly method set in
-    /// `method` when clicking this button while in the active state. Use this to create toggle buttons.
+    /// Sets a 2nd set of `args`, these arguments values are passed to the
+    /// Plotly method set in `method` when clicking this button while in the
+    /// active state. Use this to create toggle buttons.
     args2: Option<Value>,
-    /// When true, the API method is executed. When false, all other behaviors are the same and
-    /// command execution is skipped. This may be useful when hooking into, for example, the
-    /// `plotly_buttonclicked` method and executing the API command manually without losing the
-    /// benefit of the updatemenu automatically binding to the state of the plot through the
-    /// specification of `method` and `args`.
+    /// When true, the API method is executed. When false, all other behaviors
+    /// are the same and command execution is skipped. This may be useful
+    /// when hooking into, for example, the `plotly_buttonclicked` method
+    /// and executing the API command manually without losing the benefit of
+    /// the updatemenu automatically binding to the state of the plot through
+    /// the specification of `method` and `args`.
     ///
     /// Default: true
     execute: Option<bool>,
     /// Sets the text label to appear on the button.
     label: Option<String>,
-    /// Sets the Plotly method to be called on click. If the `skip` method is used, the API
-    /// updatemenu will function as normal but will perform no API calls and will not bind
-    /// automatically to state updates. This may be used to create a component interface and
-    /// attach to updatemenu events manually via JavaScript.
+    /// Sets the Plotly method to be called on click. If the `skip` method is
+    /// used, the API updatemenu will function as normal but will perform no
+    /// API calls and will not bind automatically to state updates. This may
+    /// be used to create a component interface and attach to updatemenu
+    /// events manually via JavaScript.
     method: Option<ButtonMethod>,
-    /// When used in a template, named items are created in the output figure in addition to any
-    /// items the figure already has in this array. You can modify these items in the output figure
-    /// by making your own item with `templateitemname` matching this `name` alongside your
-    /// modifications (including `visible: false` or `enabled: false` to hide it). Has no effect
-    /// outside of a template.
+    /// When used in a template, named items are created in the output figure in
+    /// addition to any items the figure already has in this array. You can
+    /// modify these items in the output figure by making your own item with
+    /// `templateitemname` matching this `name` alongside your modifications
+    /// (including `visible: false` or `enabled: false` to hide it). Has no
+    /// effect outside of a template.
     name: Option<String>,
-    /// Used to refer to a named item in this array in the template. Named items from the template
-    /// will be created even without a matching item in the input figure, but you can modify one
-    /// by making an item with `templateitemname` matching its `name`, alongside your modifications
-    /// (including `visible: false` or `enabled: false` to hide it). If there is no template or no
-    /// matching item, this item will be hidden unless you explicitly show it with `visible: true`
+    /// Used to refer to a named item in this array in the template. Named items
+    /// from the template will be created even without a matching item in
+    /// the input figure, but you can modify one by making an item with
+    /// `templateitemname` matching its `name`, alongside your modifications
+    /// (including `visible: false` or `enabled: false` to hide it). If there is
+    /// no template or no matching item, this item will be hidden unless you
+    /// explicitly show it with `visible: true`
     #[serde(rename = "templateitemname")]
     template_item_name: Option<String>,
     /// Determines whether or not this button is visible.
@@ -133,8 +142,8 @@ impl ButtonBuilder {
     }
 }
 
-/// Determines whether the buttons are accessible via a dropdown menu or whether the
-/// buttons are stacked horizontally or vertically
+/// Determines whether the buttons are accessible via a dropdown menu or whether
+/// the buttons are stacked horizontally or vertically
 ///
 /// Default: "dropdown"
 #[derive(Serialize, Debug, Clone)]
@@ -144,12 +153,11 @@ pub enum UpdateMenuType {
     Buttons,
 }
 
-/// Determines the direction in which the buttons are laid out, whether in a dropdown menu or a
-/// row/column of buttons. For `left` and `up`, the buttons will still appear in left-to-right
-/// or top-to-bottom order respectively.
+/// Determines the direction in which the buttons are laid out, whether in a
+/// dropdown menu or a row/column of buttons. For `left` and `up`, the buttons
+/// will still appear in left-to-right or top-to-bottom order respectively.
 ///
 /// Default: "down"
-///
 #[derive(Serialize, Debug, Clone)]
 #[serde(rename_all = "snake_case")]
 pub enum UpdateMenuDirection {
@@ -181,25 +189,28 @@ pub struct UpdateMenu {
     direction: Option<UpdateMenuDirection>,
     /// Sets the font of the update menu button text.
     font: Option<Font>,
-    /// When used in a template, named items are created in the output figure in addition to any
-    /// items the figure already has in this array. You can modify these items in the output figure
-    /// by making your own item with `templateitemname` matching this `name` alongside your
-    /// modifications (including `visible: false` or `enabled: false` to hide it). Has no effect
-    /// outside of a template.
+    /// When used in a template, named items are created in the output figure in
+    /// addition to any items the figure already has in this array. You can
+    /// modify these items in the output figure by making your own item with
+    /// `templateitemname` matching this `name` alongside your modifications
+    /// (including `visible: false` or `enabled: false` to hide it). Has no
+    /// effect outside of a template.
     name: Option<String>,
     /// Sets the padding around the buttons or dropdown menu.
     pad: Option<Pad>,
     /// Highlights active dropdown item or active button if true.
     #[serde(rename = "showactive")]
     show_active: Option<bool>,
-    /// Used to refer to a named item in this array in the template. Named items from the template
-    /// will be created even without a matching item in the input figure, but you can modify one by
-    /// making an item with `templateitemname` matching its `name`, alongside your modifications
-    /// (including `visible: false` or `enabled: false` to hide it). If there is no template or no
-    /// matching item, this item will be hidden unless you explicitly show it with `visible: true`.
+    /// Used to refer to a named item in this array in the template. Named items
+    /// from the template will be created even without a matching item in
+    /// the input figure, but you can modify one by making an item with
+    /// `templateitemname` matching its `name`, alongside your modifications
+    /// (including `visible: false` or `enabled: false` to hide it). If there is
+    /// no template or no matching item, this item will be hidden unless you
+    /// explicitly show it with `visible: true`.
     template_item_name: Option<String>,
-    /// Determines whether the buttons are accessible via a dropdown menu or whether the
-    /// buttons are stacked horizontally or vertically
+    /// Determines whether the buttons are accessible via a dropdown menu or
+    /// whether the buttons are stacked horizontally or vertically
     #[serde(rename = "type")]
     ty: Option<UpdateMenuType>,
     /// Determines whether or not the update menu is visible.
@@ -208,18 +219,18 @@ pub struct UpdateMenu {
     /// Default: -0.05
     /// Sets the x position (in normalized coordinates) of the update menu.
     x: Option<f64>,
-    /// Sets the update menu's horizontal position anchor. This anchor binds the `x` position to
-    /// the "left", "center" or "right" of the range selector.
-    /// Default: "right"
+    /// Sets the update menu's horizontal position anchor. This anchor binds the
+    /// `x` position to the "left", "center" or "right" of the range
+    /// selector. Default: "right"
     #[serde(rename = "xanchor")]
     x_anchor: Option<Anchor>,
     /// Type: number between or equal to -2 and 3
     /// Default: 1
     /// Sets the y position (in normalized coordinates) of the update menu.
     y: Option<f64>,
-    /// Sets the update menu's vertical position anchor This anchor binds the `y` position to the
-    /// "top", "middle" or "bottom" of the range selector.
-    /// Default: "top"
+    /// Sets the update menu's vertical position anchor This anchor binds the
+    /// `y` position to the "top", "middle" or "bottom" of the range
+    /// selector. Default: "top"
     #[serde(rename = "yanchor")]
     y_anchor: Option<Anchor>,
 }