+++ plotly/src/lib.rs
@@ -25,11 +25,11 @@ pub use configuration::Configuration;
 pub use layout::Layout;
 pub use plot::{ImageFormat, Plot, Trace};
 // Also provide easy access to modules which contain additional trace-specific types
-pub use traces::{box_plot, contour, histogram, sankey, surface};
+pub use traces::{box_plot, contour, histogram, image, mesh3d, sankey, scatter_mapbox, surface};
 // Bring the different trace types into the top-level scope
 pub use traces::{
-    Bar, BoxPlot, Candlestick, Contour, HeatMap, Histogram, Ohlc, Sankey, Scatter, Scatter3D,
-    ScatterPolar, Surface,
+    Bar, BoxPlot, Candlestick, Contour, HeatMap, Histogram, Image, Mesh3D, Ohlc, Sankey, Scatter,
+    Scatter3D, ScatterMapbox, ScatterPolar, Surface,
 };
 
 pub trait Restyle: serde::Serialize {}